requirements.txt
pandas
nltk
transformers
sentence_transformers
pinecone-client[grpc]
datasets


pytest